Implemented Authentication in backend api

+++ backend/api/azure/network.py
@@ -2,6 +2,7 @@ from flask import jsonify, request, make_response
 import json
 import requests
 import configparser
+from util.AuthenticationCheck import checkIfAuthorized
 
 config = configparser.ConfigParser()
 config.read_file(open(r'config.txt'))
@@ -11,6 +12,10 @@ networkURL = config.get('MicroServiceURL', 'azure-network')
 #CreateANetwork
 def createANetwork():
     content = request.get_json()
+
+    checkIfAuthorized(content)
+
+
     if content and content['resourcegroupname'] and content['vnetName'] and content['networkIp']:
         try:
             network = json.loads(requests.post(networkURL + '/network',json=content).text)
@@ -23,6 +28,10 @@ def createANetwork():
 #GetAllnetworks
 def getAllNetworks():
     content = request.get_json()
+
+    checkIfAuthorized(content)
+
+
     if content and content['resourcegroupname']:
         try:
             networks = json.loads(requests.patch(networkURL + '/network',json=content).text)
@@ -35,6 +44,10 @@ def getAllNetworks():
 #Get a Network
 def getANetwork(networkName):
     content = request.get_json()
+
+    checkIfAuthorized(content)
+
+
     if networkName and content and content['resourcegroupname']:
         try:
             network = json.loads(requests.patch(networkURL + '/network/'+networkName,json=content).text)
@@ -47,6 +60,10 @@ def getANetwork(networkName):
 #Delete Network
 def deleteNetwork():
     content = request.get_json()
+
+    checkIfAuthorized(content)
+
+
     if content and content['resourcegroupname'] and content['vnetName']:
         try:
             network = json.loads(requests.delete(networkURL + '/network',json=content).text)
@@ -59,6 +76,10 @@ def deleteNetwork():
 #Createsubnet
 def createOrUpdateSubnet():
     content = request.get_json()
+
+    checkIfAuthorized(content)
+
+
     if content and content['resourcegroupname'] and content['vnetName'] and content['subnetName'] and content['subnetIp']:
         try:
             subnet = json.loads(requests.post(networkURL + '/subnet',json=content).text)
@@ -71,6 +92,10 @@ def createOrUpdateSubnet():
 #Create NIC
 def CreateOrUpdateNic():
     content = request.get_json()
+
+    checkIfAuthorized(content)
+
+
     if content and content['resourcegroupname'] and content['nicname'] and content['ipconfigname'] and content['subnetid'] and content['vnetName']:
         try:
             nic = json.loads(requests.post(networkURL + '/nic',json=content).text)
@@ -83,6 +108,10 @@ def CreateOrUpdateNic():
 #get all NIC
 def getALLNIC():
     content = request.get_json()
+
+    checkIfAuthorized(content)
+
+    
     if content and content['resourcegroupname']:
         try:
             nic = json.loads(requests.patch(networkURL + '/nic',json=content).text)
